+++ examples/basic-iframe/.tina/config.js
@@ -89,6 +89,8 @@ export default defineConfig({
         path: 'content/page',
         label: 'Page',
         format: 'mdx',
+        frontmatterFormat: 'toml',
+        frontmatterDelimiters: ['+++', '+++'],
         fields: [
           {
             label: 'Title',