+++ packages/@tinacms/app/appFiles/src/lib/machines/document-machine.ts
@@ -208,7 +208,7 @@ export const documentMachine =
           const onSubmit = async (payload: Record<string, unknown>) => {
             try {
               const mutationString = `#graphql
-              mutation UpdateDocument($collection: String!, $relativePath: String!, $params: DocumentMutation!) {
+              mutation UpdateDocument($collection: String!, $relativePath: String!, $params: DocumentUpdateMutation!) {
                 updateDocument(collection: $collection, relativePath: $relativePath, params: $params) {
                   __typename
                 }