+++ packages/@tinacms/cli/src/cmds/baseCmds.ts
@@ -45,6 +45,11 @@ const startServerPortOption = {
   name: '--port <port>',
   description: 'Specify a port to run the server on. (default 4001)',
 }
+const rootPathOption = {
+  name: '--rootPath <rootPath>',
+  description:
+    'Specify the root directory to run the CLI from (defaults to current working directory)',
+}
 const experimentalDatalayer = {
   name: '--experimentalData',
   description: 'Build the server with additional data querying capabilities',
@@ -164,6 +169,7 @@ export const baseCmds: Command[] = [
       noTelemetryOption,
       watchFileOption,
       verboseOption,
+      rootPathOption,
     ],
     action: (options) =>
       chain(
@@ -189,6 +195,7 @@ export const baseCmds: Command[] = [
       verboseOption,
       developmentOption,
       localOption,
+      rootPathOption,
     ],
     action: (options) =>
       chain(
@@ -208,6 +215,7 @@ export const baseCmds: Command[] = [
   {
     command: INIT,
     options: [
+      rootPathOption,
       experimentalDatalayer,
       isomorphicGitBridge,
       noTelemetryOption,
@@ -220,6 +228,7 @@ export const baseCmds: Command[] = [
   },
   {
     options: [
+      rootPathOption,
       cleanOption,
       useDefaultValuesOption,
       noTelemetryOption,