+++ packages/@tinacms/cli/src/cmds/statusChecks/checkClientInformation.ts
@@ -16,39 +16,32 @@ import type { Bridge } from '@tinacms/datalayer'
 import type { Database } from '@tinacms/graphql'
 import type { TinaCloudSchema } from '@tinacms/schema-tools'
 import { ConfigBuilder } from '../../buildTina'
+import { parseURL } from './waitForIndexing'
 
 //  This was taken from packages/tinacms/src/unifiedClient/index.ts
 // TODO: maybe move this to a shared util package?
 
-async function request<DataType extends Record<string, any> = any>(args: {
+async function request(args: {
   url: string
   token: string
-  variables?: Record<string, any>
-  query: string
-}): Promise<{ data: DataType; query: string }> {
-  let data: DataType = {} as DataType
+}): Promise<{ status: string; timestamp: number }> {
   const headers = new Headers()
   if (args.token) {
     headers.append('X-API-KEY', args.token)
   }
   headers.append('Content-Type', 'application/json')
 
-  const bodyString = JSON.stringify({
-    query: args.query,
-    variables: args?.variables || {},
-  })
   const url = args?.url
 
   const res = await fetch(url, {
-    method: 'POST',
+    method: 'GET',
     headers,
-    body: bodyString,
     redirect: 'follow',
   })
   const json = await res.json()
   if (!res.ok) {
     let additionalInfo = ''
-    if (res.status === 401) {
+    if (res.status === 401 || res.status === 403) {
       additionalInfo =
         'Please check that your client ID, URL and read only token are configured properly.'
     }
@@ -69,8 +62,8 @@ async function request<DataType extends Record<string, any> = any>(args: {
     )
   }
   return {
-    data: json?.data as DataType,
-    query: args.query,
+    status: json?.status,
+    timestamp: json?.timestamp,
   }
 }
 
@@ -94,19 +87,14 @@ export const checkClientInfo = async (
 
   const config = ctx.schema?.config
   const token = config.token
-  const url = ctx.apiUrl
-
+  const { clientId, branch, host } = parseURL(ctx.apiUrl)
+  const url = `https://${host}/db/${clientId}/status/${branch}`
   const bar = new Progress('Checking clientId, token and branch. :prog', 1)
 
   try {
     await request({
       token,
       url,
-      query: `query {
-        collections {
-          name
-        }
-      }`,
     })
     bar.tick({
       prog: '✅',