+++ packages/@tinacms/graphql/src/database/index.ts
@@ -459,7 +459,11 @@ export class Database {
       const stringifiedFile = stringifyFile(
         payload,
         extension,
-        templateInfo.type === 'union'
+        templateInfo.type === 'union',
+        {
+          frontmatterFormat: collection?.frontmatterFormat,
+          frontmatterDelimiters: collection?.frontmatterDelimiters,
+        }
       )
       return {
         stringifiedFile,
@@ -1091,8 +1095,14 @@ const _indexContent = async (
   await sequential(documentPaths, async (filepath) => {
     try {
       const dataString = await database.bridge.get(normalizePath(filepath))
-      const data = parseFile(dataString, path.extname(filepath), (yup) =>
-        yup.object({})
+      const data = parseFile(
+        dataString,
+        path.extname(filepath),
+        (yup) => yup.object({}),
+        {
+          frontmatterDelimiters: collection?.frontmatterDelimiters,
+          frontmatterFormat: collection?.frontmatterFormat,
+        }
       )
       const normalizedPath = normalizePath(filepath)
       await enqueueOps([