+++ packages/@tinacms/graphql/src/resolver/index.ts
@@ -407,6 +407,7 @@ export class Resolver {
     isDeletion,
     isAddPendingDocument,
     isCollectionSpecific,
+    isUpdateName,
   }: {
     args: unknown
     collection?: string
@@ -415,6 +416,7 @@ export class Resolver {
     isDeletion?: boolean
     isAddPendingDocument?: boolean
     isCollectionSpecific?: boolean
+    isUpdateName?: boolean
   }) => {
     /**
      * `collectionName` is passed in:
@@ -469,15 +471,43 @@ export class Resolver {
           isAddPendingDocument,
         })
       }
-      if (isDeletion) {
-        if (!alreadyExists) {
+      // if we are deleting a document or updating its name we should check if it exists
+      if (!alreadyExists) {
+        if (isDeletion) {
           throw new Error(
             `Unable to delete document, ${realPath} does not exist`
           )
         }
+        if (isUpdateName) {
+          throw new Error(
+            `Unable to update document, ${realPath} does not exist`
+          )
+        }
+      }
+      if (isDeletion) {
         const doc = await this.getDocument(realPath)
         await this.deleteDocument(realPath)
         return doc
+      }
+      if (isUpdateName) {
+        // Must provide a new relative path in the params
+        assertShape<{ params: string }>(args, (yup) =>
+          yup.object({ params: yup.object().required() })
+        )
+        assertShape<{ relativePath: string }>(args?.params, (yup) =>
+          yup.object({ relativePath: yup.string().required() })
+        )
+        // Get the real document
+        const doc = await this.getDocument(realPath)
+        const newRealPath = path.join(
+          collection?.path,
+          args.params.relativePath
+        )
+        // Update the document
+        await this.database.put(newRealPath, doc._rawData, collection.name)
+        // Delete the old document
+        await this.deleteDocument(realPath)
+        return this.getDocument(newRealPath)
       }
       /**
        * updateDocument, update<Collection>Document
@@ -675,6 +705,7 @@ export class Resolver {
     Object.entries(fieldParams).forEach(([fieldName, fieldValue]) => {
       if (Array.isArray(fieldValue)) {
         if (fieldValue.length === 0) {
+          accum[fieldName] = []
           return
         }
       }