+++ packages/@tinacms/mdx/src/tests/autotest/_config.ts
@@ -156,8 +156,8 @@ export const field: RichTypeInner = {
       ],
     },
     {
-      name: 'Shortcode1',
-      label: 'Shortcode 1',
+      name: 'rimg',
+      label: 'rimg',
       inline: true,
       match: {
         start: '{{<',
@@ -165,35 +165,79 @@ export const field: RichTypeInner = {
       },
       fields: [
         {
-          name: 'text',
-          label: 'Text',
+          name: 'src',
+          label: 'Src',
           type: 'string',
           required: true,
           isTitle: true,
-          ui: {
-            component: 'textarea',
-          },
         },
       ],
     },
     {
-      name: 'Shortcode2',
-      label: 'Shortcode 2',
+      name: 'adPanel',
+      label: 'Ad Panel',
       inline: true,
       match: {
         start: '{{%',
         end: '%}}',
+        name: 'ad-panel-leaderboard',
       },
       fields: [
         {
-          name: 'text',
+          name: '_value',
           required: true,
           isTitle: true,
-          label: 'Text',
+          label: 'Value',
           type: 'string',
-          ui: {
-            component: 'textarea',
-          },
+        },
+      ],
+    },
+    {
+      name: 'featurePanel',
+      label: 'Feature Panel',
+      inline: true,
+      match: {
+        start: '{{%',
+        end: '%}}',
+        name: 'feature-panel',
+      },
+      fields: [
+        {
+          name: '_value',
+          required: true,
+          isTitle: true,
+          label: 'Value',
+          type: 'string',
+        },
+      ],
+    },
+    {
+      name: 'signature',
+      label: 'Signature',
+      match: {
+        start: '{{<',
+        end: '>}}',
+      },
+      fields: [
+        {
+          name: 'foo',
+          label: 'foo label',
+          type: 'string',
+        },
+      ],
+    },
+    {
+      name: 'unkeyedSignature',
+      label: 'Unkeyed Signature',
+      match: {
+        start: '{{<',
+        end: '>}}',
+      },
+      fields: [
+        {
+          name: '_value',
+          label: 'Value',
+          type: 'string',
         },
       ],
     },