+++ packages/@tinacms/schema-tools/src/validate/fields.ts
@@ -133,6 +133,7 @@ export const TinaFieldZod: z.ZodType<TinaFieldInner<false>> = z.lazy(() => {
         .object({
           start: z.string(),
           end: z.string(),
+          name: z.string().optional(),
         })
         .optional(),
     })